+++ crates/matrix-sdk/src/encryption/tasks.rs
@@ -16,13 +16,10 @@ use std::{collections::BTreeMap, sync::Arc, time::Duration};
 
 use futures_core::Stream;
 use futures_util::{StreamExt, pin_mut};
+use matrix_sdk_base::crypto::store::types::{RoomKeyBundleInfo, RoomPendingKeyBundleDetails};
 #[cfg(feature = "experimental-encrypted-state-events")]
 use matrix_sdk_base::crypto::types::events::room::encrypted::{
     EncryptedEvent, RoomEventEncryptionScheme,
-};
-use matrix_sdk_base::{
-    RoomState,
-    crypto::store::types::{RoomKeyBundleInfo, RoomPendingKeyBundleDetails},
 };
 use matrix_sdk_common::failures_cache::FailuresCache;
 #[cfg(not(feature = "experimental-encrypted-state-events"))]
@@ -449,6 +446,10 @@ impl BundleReceiverTask {
     async fn listen_task(client: WeakClient, stream: impl Stream<Item = RoomKeyBundleInfo>) {
         pin_mut!(stream);
 
+        // Before we listen for new bundles, we check if we have existing ones
+        // unimported.
+        Self::try_import_stored_bundles(&client).await;
+
         // TODO: Listening to this stream is not enough for iOS due to the NSE killing
         // our OlmMachine and thus also this stream. We need to add an event handler
         // that will listen for the bundle event. To be able to add an event handler,
@@ -466,6 +467,36 @@ impl BundleReceiverTask {
             };
 
             Self::handle_bundle(&room, &bundle_info).await;
+        }
+    }
+
+    /// Retrieves a list of all rooms pending key bundles, then cross-references
+    /// this with bundles held in the crypto store. If all conditions outlined
+    /// in [`shared_room_history::maybe_accept_key_bundle`], then the bundle
+    /// will be imported.
+    async fn try_import_stored_bundles(client: &WeakClient) {
+        let Some(client) = client.get() else {
+            return;
+        };
+
+        let olm_machine = client.olm_machine().await;
+        let Some(olm_machine) = olm_machine.as_ref() else {
+            return;
+        };
+        let Ok(room_details) = olm_machine.store().get_all_rooms_pending_key_bundles().await else {
+            return;
+        };
+
+        for RoomPendingKeyBundleDetails { room_id, inviter, .. } in &room_details {
+            let Some(room) = client.get_room(room_id) else {
+                continue;
+            };
+            let Ok(Some(bundle)) =
+                olm_machine.store().get_received_room_key_bundle_data(room_id, inviter).await
+            else {
+                continue;
+            };
+            Self::handle_bundle(&room, &(&bundle).into()).await;
         }
     }
 
@@ -488,7 +519,7 @@ impl BundleReceiverTask {
     /// thread will process it.
     #[instrument(skip(room), fields(room_id = %room.room_id()))]
     async fn handle_bundle(room: &Room, bundle_info: &RoomKeyBundleInfo) {
-        if Self::should_accept_bundle(room, bundle_info).await {
+        if shared_room_history::should_accept_key_bundle(room, bundle_info).await {
             info!("Accepting a late key bundle.");
 
             if let Err(e) =
@@ -501,50 +532,23 @@ impl BundleReceiverTask {
         }
     }
 
-    async fn should_accept_bundle(room: &Room, bundle_info: &RoomKeyBundleInfo) -> bool {
-        // We accept historic room key bundles up to one day after we have accepted an
-        // invite.
-        const DAY: Duration = Duration::from_secs(24 * 60 * 60);
-
-        // If we don't have any invite acceptance details, then this client wasn't the
-        // one that accepted the invite.
-        let Ok(Some(RoomPendingKeyBundleDetails { invite_accepted_at, inviter, .. })) =
-            room.client.base_client().get_pending_key_bundle_details_for_room(room.room_id()).await
-        else {
-            debug!("Not accepting key bundle as there are no recorded invite acceptance details");
-            return false;
-        };
-
-        let state = room.state();
-        let elapsed_since_join = invite_accepted_at.to_system_time().and_then(|t| t.elapsed().ok());
-        let bundle_sender = &bundle_info.sender;
-
-        match (state, elapsed_since_join) {
-            (RoomState::Joined, Some(elapsed_since_join)) => {
-                elapsed_since_join < DAY && bundle_sender == &inviter
-            }
-            (RoomState::Joined, None) => false,
-            (RoomState::Left | RoomState::Invited | RoomState::Knocked | RoomState::Banned, _) => {
-                false
-            }
-        }
+    #[cfg(any(feature = "testing", test))]
+    pub(crate) fn abort(&self) {
+        self._handle.abort();
     }
 }
 
 #[cfg(all(test, not(target_family = "wasm")))]
 mod test {
-    use matrix_sdk_test::{
-        InvitedRoomBuilder, JoinedRoomBuilder, async_test, event_factory::EventFactory,
-    };
+    use matrix_sdk_test::async_test;
     #[cfg(not(feature = "experimental-encrypted-state-events"))]
     use ruma::events::room::encrypted::OriginalSyncRoomEncryptedEvent;
-    use ruma::{event_id, room_id, user_id};
+    use ruma::{event_id, room_id};
     use serde_json::json;
-    use vodozemac::Curve25519PublicKey;
     use wiremock::MockServer;
 
     use super::*;
-    use crate::test_utils::{logged_in_client, mocks::MatrixMockServer};
+    use crate::test_utils::logged_in_client;
 
     // Test that, if backups are not enabled, we don't incorrectly mark a room key
     // as downloaded.
@@ -605,90 +609,5 @@ mod test {
                 "Backups are not enabled, we should not mark any room keys as downloaded."
             )
         }
-    }
-
-    /// Test that ensures that we only accept a bundle if a certain set of
-    /// conditions is met.
-    #[async_test]
-    async fn test_should_accept_bundle() {
-        let server = MatrixMockServer::new().await;
-
-        let alice_user_id = user_id!("@alice:localhost");
-        let bob_user_id = user_id!("@bob:localhost");
-        let joined_room_id = room_id!("!joined:localhost");
-        let invited_rom_id = room_id!("!invited:localhost");
-
-        let client = server
-            .client_builder()
-            .logged_in_with_token("ABCD".to_owned(), alice_user_id.into(), "DEVICEID".into())
-            .build()
-            .await;
-
-        let event_factory = EventFactory::new().room(invited_rom_id);
-        let bob_member_event = event_factory.member(bob_user_id);
-        let alice_member_event = event_factory.member(bob_user_id).invited(alice_user_id);
-
-        server
-            .mock_sync()
-            .ok_and_run(&client, |builder| {
-                builder.add_joined_room(JoinedRoomBuilder::new(joined_room_id)).add_invited_room(
-                    InvitedRoomBuilder::new(invited_rom_id)
-                        .add_state_event(bob_member_event)
-                        .add_state_event(alice_member_event),
-                );
-            })
-            .await;
-
-        let room =
-            client.get_room(joined_room_id).expect("We should have access to our joined room now");
-
-        assert!(
-            client
-                .base_client()
-                .get_pending_key_bundle_details_for_room(room.room_id())
-                .await
-                .unwrap()
-                .is_none(),
-            "We shouldn't have any invite acceptance details if we didn't join the room on this Client"
-        );
-
-        let bundle_info = RoomKeyBundleInfo {
-            sender: bob_user_id.to_owned(),
-            sender_key: Curve25519PublicKey::from_bytes([0u8; 32]),
-            room_id: joined_room_id.to_owned(),
-        };
-
-        assert!(
-            !BundleReceiverTask::should_accept_bundle(&room, &bundle_info).await,
-            "We should not accept a bundle if we did not join the room from this Client"
-        );
-
-        let invited_room =
-            client.get_room(invited_rom_id).expect("We should have access to our invited room now");
-
-        assert!(
-            !BundleReceiverTask::should_accept_bundle(&invited_room, &bundle_info).await,
-            "We should not accept a bundle if we didn't join the room."
-        );
-
-        server.mock_room_join(invited_rom_id).ok().mock_once().mount().await;
-
-        let room = client
-            .join_room_by_id(invited_rom_id)
-            .await
-            .expect("We should be able to join the invited room");
-
-        let details = client
-            .base_client()
-            .get_pending_key_bundle_details_for_room(room.room_id())
-            .await
-            .unwrap()
-            .expect("We should have stored the invite acceptance details");
-        assert_eq!(details.inviter, bob_user_id, "We should have recorded that Bob has invited us");
-
-        assert!(
-            BundleReceiverTask::should_accept_bundle(&room, &bundle_info).await,
-            "We should accept a bundle if we just joined the room and did so from this very Client object"
-        );
     }
 }